+++ client/src/pages/profile/styles.scss
@@ -61,6 +61,8 @@
   text-align: left;
 }
 
-/* .checkbox {
-  padding-left: 20% !important;
-} */
+.profile-page {
+  .checkbox {
+    padding-left: 20% !important;
+  }
+}
