+++ client/src/utils/neighborhoods.ts
@@ -1,0 +1,251 @@
+const neighborhoods = [
+  'Albany Park',
+  'Altgeld Gardens',
+  'Andersonville',
+  'Arcadia Terrace',
+  'Archer Heights',
+  'Armour Square',
+  'Ashburn',
+  'Ashburn Estates',
+  'Auburn Gresham',
+  'Avalon Park',
+  'Avondale',
+  'Avondale Gardens',
+  'Back of the Yards',
+  'Belmont Central',
+  'Belmont Gardens',
+  'Belmont Heights',
+  'Belmont Terrace',
+  'Beverly',
+  'Beverly View',
+  'Beverly Woods',
+  'Big Oaks',
+  'Northalsted "Boystown"',
+  'Bowmanville',
+  'Brainerd',
+  'Brickyard',
+  'Bridgeport',
+  'Brighton Park',
+  'Bronzeville',
+  'Bucktown',
+  'Budlong Woods',
+  'Buena Park',
+  'Burnside',
+  'Cabrini–Green',
+  'Calumet Heights',
+  'Canaryville',
+  'Central Station',
+  'Chatham',
+  'Chicago Lawn',
+  'Chinatown',
+  'Chrysler Village',
+  'Clarendon Park',
+  'Clearing East',
+  'Clearing West',
+  'Cottage Grove Heights',
+  'Cragin',
+  'Crestline',
+  'Dearborn Homes',
+  'Dearborn Park',
+  'Douglas Park',
+  'Dunning',
+  'East Beverly',
+  'East Chatham',
+  'East Garfield Park',
+  'East Hyde Park',
+  'East Pilsen',
+  'East Side',
+  'East Village',
+  'Eden Green',
+  'Edgebrook',
+  'Edgewater',
+  'Edgewater Beach',
+  'Edgewater Glen',
+  'Edison Park',
+  'Englewood',
+  'Fernwood',
+  'Fifth City',
+  'Ford City',
+  'Forest Glen',
+  'Fuller Park',
+  'Fulton River District',
+  'Gage Park',
+  'Galewood',
+  'The Gap',
+  'Garfield Ridge',
+  'Gladstone Park',
+  'Gold Coast',
+  'Golden Gate',
+  'Goose Island',
+  'Graceland West',
+  'Grand Boulevard',
+  'Grand Crossing',
+  'Greater Grand Crossing',
+  'Greektown',
+  'Gresham',
+  'Groveland Park',
+  'Hamilton Park',
+  'Hanson Park',
+  'Heart of Chicago',
+  'Hegewisch',
+  'Hermosa',
+  'Hollywood Park',
+  'Homan Square',
+  'Humboldt Park',
+  'Hyde Park',
+  'Illinois Medical District',
+  'Irving Park',
+  'Irving Woods',
+  'The Island',
+  'Jackowo',
+  'Jackson Park Highlands',
+  'Jefferson Park',
+  'K-Town',
+  'Kelvyn Park',
+  'Kennedy Park',
+  'Kensington',
+  'Kenwood',
+  'Kilbourn Park',
+  'Kosciuszko Park',
+  'Lake Meadows',
+  'Lake View',
+  'Lakeview (East)',
+  'Lakewood / Balmoral',
+  'LeClaire Courts',
+  'Legends South (Robert Taylor Homes)',
+  'Lilydale',
+  'Lincoln Park',
+  'Lincoln Square',
+  'Lithuanian Plaza',
+  'Little Italy',
+  'Little Village',
+  'Logan Square',
+  'Longwood Manor',
+  'The Loop',
+  'Lower West Side',
+  'Loyola',
+  'Magnificent Mile',
+  'Margate Park',
+  'Marquette Park',
+  'Marshall Square',
+  'Marynook',
+  'Mayfair',
+  'McKinley Park',
+  'Merchant Park',
+  'Montclare',
+  'Morgan Park',
+  'Mount Greenwood',
+  'Museum Campus',
+  'New Eastside',
+  'Near North Side',
+  'Near West Side',
+  'New Chinatown',
+  'New City',
+  'Noble Square',
+  'North Austin',
+  'North Center',
+  'North Halsted',
+  'North Kenwood',
+  'North Lawndale',
+  'North Mayfair',
+  'North Park',
+  'Nortown',
+  'Norwood Park East',
+  'Norwood Park West',
+  'Oakland',
+  "O'Hare",
+  'Old Edgebrook',
+  'Old Irving Park',
+  'Old Norwood',
+  'Old Town',
+  'Old Town Triangle',
+  'Oriole Park',
+  'Palmer Square',
+  'Park Manor',
+  'Park West',
+  'Parkview',
+  'Peterson Park',
+  'Pill Hill',
+  'Pilsen',
+  'Polish Downtown',
+  'Polish Village',
+  'Portage Park',
+  'Prairie Avenue Historic District',
+  'Prairie Shores',
+  'Princeton Park',
+  "Printer's Row",
+  'Pulaski Park',
+  'Pullman',
+  'Ranch Triangle',
+  'Ravenswood',
+  'Ravenswood Gardens',
+  'Ravenswood Manor',
+  'River North',
+  'River West',
+  "River's Edge",
+  'Riverdale',
+  'Rogers Park',
+  'Roscoe Village',
+  'Rosehill',
+  'Roseland',
+  'Rosemoor',
+  "St. Ben's",
+  'Sauganash',
+  'Schorsch Forest View',
+  'Schorsch Village',
+  'Scottsdale',
+  'Sheffield Neighbors',
+  'Sheridan Park',
+  'Sheridan Station Corridor',
+  'Sleepy Hollow',
+  'Smith Park',
+  'South Austin',
+  'South Chicago',
+  'South Commons',
+  'South Deering',
+  'South East Ravenswood',
+  'South Edgebrook',
+  'South Lawndale',
+  'South Loop',
+  'South Shore',
+  'Stateway Gardens',
+  'Stony Island Park',
+  'Streeterville',
+  "Talley's Corner",
+  'Tri-Taylor',
+  'Ukrainian Village',
+  'Union Ridge',
+  'University Village',
+  'Uptown',
+  'The Villa',
+  'Vittum Park',
+  'Wacławowo',
+  'Washington Heights',
+  'Washington Park',
+  'Wentworth Gardens',
+  'West Beverly',
+  'West Chatham',
+  'West Chesterfield',
+  'West DePaul',
+  'West Elsdon',
+  'West Englewood',
+  'West Garfield Park',
+  'West Humboldt Park',
+  'Lakeview (West)',
+  'West Lawn',
+  'West Loop',
+  'West Morgan Park',
+  'West Pullman',
+  'West Ridge',
+  'West Rogers Park',
+  'West Town',
+  'West Woodlawn',
+  'Wicker Park',
+  'Wildwood',
+  'Woodlawn',
+  'Wrightwood',
+  'Wrightwood Neighbors',
+  'Wrigleyville',
+];
+
+export default neighborhoods;
